Compile UdmDll (without MGA) under x64
git-svn-id: https://svn.isis.vanderbilt.edu/MoBIES/UDM/trunk@3860 c0adbae9-c137-0410-b742-ec4b2d2ceded

+++ src/UdmDom/UdmDom.cpp
@@ -753,7 +753,7 @@ namespace UdmDom
 		{
 			UDM_ASSERT(src != NULL);
 
-			int ret_length = XMLString::stringLen(src) + 1 + data.length() + 1;
+			size_t ret_length = XMLString::stringLen(src) + 1 + data.length() + 1;
 			XMLCh *retval = (XMLCh*) XMLPlatformUtils::fgMemoryManager->allocate(ret_length * sizeof(XMLCh));
 			XMLString::copyString(retval, src);
 			XMLString::catString(retval, gXML_space);
